+++ lib/default/TasmotaSerial-3.5.0/src/TasmotaSerial.cpp
@@ -48,12 +48,14 @@ TasmotaSerial::TasmotaSerial(int receive_pin, int transmit_pin, int hardware_fal
   m_valid = false;
   m_hardserial = false;
   m_hardswap = false;
+  m_overflow = false;
   m_stop_bits = 1;
   m_nwmode = nwmode;
   serial_buffer_size = buffer_size;
   m_rx_pin = receive_pin;
   m_tx_pin = transmit_pin;
-  m_in_pos = m_out_pos = 0;
+  m_in_pos = 0;
+  m_out_pos = 0;
 #ifdef ESP8266
   if (!((isValidGPIOpin(receive_pin)) && (isValidGPIOpin(transmit_pin) || transmit_pin == 16))) {
     return;
@@ -162,11 +164,11 @@ size_t TasmotaSerial::setRxBufferSize(size_t size) {
   if (size != serial_buffer_size) {
     if (m_hardserial) {
       if (size > 256) {      // Default hardware serial Rx buffer size
-  #ifdef ESP8266
+#ifdef ESP8266
         serial_buffer_size = size;
         Serial.setRxBufferSize(serial_buffer_size);
-  #endif  // ESP8266
-  #ifdef ESP32
+#endif  // ESP8266
+#ifdef ESP32
         if (TSerial) {
           // RX Buffer can't be resized when Serial is already running
           serial_buffer_size = size;
@@ -176,7 +178,7 @@ size_t TasmotaSerial::setRxBufferSize(size_t size) {
           TSerial->setRxBufferSize(serial_buffer_size);
           Esp32Begin();
         }
-  #endif  // ESP32
+#endif  // ESP32
       }
     }
     else if (m_buffer) {
@@ -267,6 +269,21 @@ bool TasmotaSerial::hardwareSerial(void) {
 #endif  // ESP32
 }
 
+bool TasmotaSerial::overflow(void) {
+  if (m_hardserial) {
+#ifdef ESP8266
+    return Serial.hasOverrun();  // Returns then clear overrun flag
+#endif  // ESP8266
+#ifdef ESP32
+    return false;  // Not implemented
+#endif  // ESP32
+  } else {
+    bool res = m_overflow;
+    m_overflow = false;
+    return res;
+  }
+}
+
 void TasmotaSerial::flush(void) {
   if (m_hardserial) {
 #ifdef ESP8266
@@ -277,7 +294,8 @@ void TasmotaSerial::flush(void) {
     while (TSerial->available()) { TSerial->read(); }
 #endif  // ESP32
   } else {
-    m_in_pos = m_out_pos = 0;
+    m_in_pos = 0;
+    m_out_pos = 0;
   }
 }
 
@@ -322,7 +340,7 @@ size_t TasmotaSerial::read(char* buffer, size_t size) {
   } else {
     if ((-1 == m_rx_pin) || (m_in_pos == m_out_pos)) { return 0; }
     size_t count = 0;
-    for( ; size && (m_in_pos == m_out_pos) ; --size, ++count) {
+    for( ; size && (m_in_pos != m_out_pos) ; --size, ++count) {
       *buffer++ = m_buffer[m_out_pos];
       m_out_pos = (m_out_pos +1) % serial_buffer_size;
     }
@@ -341,6 +359,11 @@ int TasmotaSerial::available(void) {
   } else {
     int avail = m_in_pos - m_out_pos;
     if (avail < 0) avail += serial_buffer_size;
+
+//    if (!avail) {
+//      optimistic_yield(10000);
+//    }
+
     return avail;
   }
 }
@@ -407,13 +430,14 @@ size_t TasmotaSerial::write(uint8_t b) {
 }
 
 void IRAM_ATTR TasmotaSerial::rxRead(void) {
+  uint32_t m_out_pos_fixed = m_out_pos;
   if (!m_nwmode) {
     int32_t loop_read = m_very_high_speed ? serial_buffer_size : 1;
     // Advance the starting point for the samples but compensate for the
     // initial delay which occurs before the interrupt is delivered
     uint32_t wait = m_bit_start_time;
     uint32_t start = ESP.getCycleCount();
-    while (loop_read-- > 0) {    // try to receveive all consecutive bytes in a raw
+    while (loop_read-- > 0) {    // try to receveive all consecutive bytes in a row
       uint32_t rec = 0;
       for (uint32_t i = 0; i < 8; i++) {
         TM_SERIAL_WAIT_RCV;
@@ -421,10 +445,14 @@ void IRAM_ATTR TasmotaSerial::rxRead(void) {
         if (digitalRead(m_rx_pin)) rec |= 0x80;
       }
       // Store the received value in the buffer unless we have an overflow
-      uint32_t next = (m_in_pos+1) % serial_buffer_size;
-      if (next != (int)m_out_pos) {
+      uint32_t next = (m_in_pos + 1) % serial_buffer_size;
+      if (next != m_out_pos_fixed) {
         m_buffer[m_in_pos] = rec;
         m_in_pos = next;
+      } else {
+        // Buffer overrun - exit
+        m_overflow = true;
+        loop_read = 0;
       }
 
       TM_SERIAL_WAIT_RCV_LOOP;    // wait for stop bit
@@ -486,7 +514,7 @@ void IRAM_ATTR TasmotaSerial::rxRead(void) {
         }
         //stobyte(0,ssp->ss_byte>>1);
         uint32_t next = (m_in_pos + 1) % serial_buffer_size;
-        if (next != (uint32_t)m_out_pos) {
+        if (next != m_out_pos_fixed) {
           m_buffer[m_in_pos] = ss_byte >> 1;
           m_in_pos = next;
         }
@@ -500,7 +528,7 @@ void IRAM_ATTR TasmotaSerial::rxRead(void) {
         // bit zero was 0,
         //stobyte(0,ssp->ss_byte>>1);
         uint32_t next = (m_in_pos + 1) % serial_buffer_size;
-        if (next != (uint32_t)m_out_pos) {
+        if (next != m_out_pos_fixed) {
           m_buffer[m_in_pos] = ss_byte >> 1;
           m_in_pos = next;
         }