+++ CHANGELOG.md
@@ -3,23 +3,39 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased] - Development
 
-## [12.2.0.5]
+## [12.2.0.6]
+### Added
+
+### Breaking Changed
+
+### Changed
+
+### Fixed
+- TasmotaSerial ``read(buffer, size)`` regression from v9.3.0
+
+### Removed
+
+## [12.2.0.5] 20221129
 ### Added
 - ESP32 DS18x20 parasitic power usage when defining W1_PARASITE_POWER (#17112)
 - Optional define ``SERIAL_BRIDGE_BUFFER_SIZE`` to set Serial Bridge internal buffer size (Default ESP8266 = 256, ESP32 = 800)
 - Command ``SSerialBuffer 256..SERIAL_BRIDGE_BUFFER_SIZE`` to change serial bridge rx buffer size (#17120)
 - Command ``SetOption35 0..255`` to skip number of received messages in Serial Bridge (default 0) (#17140)
-
-### Breaking Changed
+- Teleinfo TEMPO (BBR) contract (#17160)
+- Support for HLK-LD2410 24GHz smart wave motion sensor
+- Berry ``mdns`` module (#17202)
+- IPv6 preview for ESP32, also working for ESP8266
 
 ### Changed
 - Serial Bridge default internal serial rx buffer size from 64 to 256 (#17120)
 - Accept filename extensions to GUI file upload input fields (#16875)
+- AC PWM dimmer lineair power distribution (#17177)
 
 ### Fixed
 - ModbusBridge baudrates over 76500 baud (#17106)
 
 ### Removed
+- Accept filename extensions to GUI file upload input fields as not functional in some browsers (#16875)
 
 ## [12.2.0.4] 20221117
 ### Added